[#347] Cleaning up fixtures in the main conftest
