+++ CHANGELOG.md
@@ -7,12 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - YYYY-MM-DD
 ### Added
 ### Changed
+- Cleaning up the main conftest [@Michal-Kolomanski]
+### Fixed
+### Removed
+
+## [3.2.1] - 2022-05-05
+### Added
+### Changed
 - Python version: 3.10 [@Michal-Kolomanski]
 - Updating and cleaning up packages [@Michal-Kolomanski]
 ### Fixed
 ### Removed
 
-## [3.2] - 2022-04-26
+## [3.2.0] - 2022-04-26
 ### Added
 - Integrate elastic_search results from the Marketplace [@Michal-Kolomanski]
 - JMS Subscriber for user actions and flask cli for it [@michal-szostak]
